Split webhook events types
