+++ packages/types/src/index.ts
@@ -1,4 +1,5 @@
 export * from './discord.js'
 export * from './discord/reference.js'
+export * from './discord/webhookEvents.js'
 export * from './discordeno.js'
 export * from './shared.js'
